common/common-nut_version.c, common/common.c: evict UPS_VERSION and its consumers into a separate compilation unit [#2097]
