+++ NEWS.adoc
@@ -41,6 +41,10 @@ https://github.com/networkupstools/nut/milestone/9
 
  - (expected) Bug fixes for fallout possible due to "fightwarn" effort in 2.8.0+
 
+ - Refactored NUT "common" sources to reference `nut_version.h` macros from
+   a smaller C source file, to minimize the compilation unit size impacted by
+   development iterations. [issue #2097]
+
 Release notes for NUT 2.8.3 - what's new since 2.8.2
 ----------------------------------------------------
